+++ CHANGES.md
@@ -9,3 +9,13 @@ Pyrengine is based on my previous project Pyrone rewritten in modern flask from 
 
 Key differences: no more users support except site admin, at least for now. So no registration,
 no external logins via external services.
+
+# 1.0.1
+
+Released: 2023-09-xx
+
+* backups features added: create backup, delete backup
+* fixed SELECT html element render, added missing "selected" attribute
+* fixed broken translations
+* added site search widget
+* added favicon instructions